Check for file first

+++ scripts/make_run_server.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 export PORT=8003
 
-rm scripts/run_server.sh
+if [ -f scripts/run_server.sh ] ; then
+    rm scripts/run_server.sh
+fi
 echo "#!/bin/bash" >> scripts/run_server.sh
 echo "export PATH=$PATH" >> scripts/run_server.sh
 echo "python manage.py migrate" >> scripts/run_server.sh
